Fixing StartupLoader tests
