+++ src/Tests/Gate.Builder.Tests/Gate.Builder.Tests.csproj
@@ -34,8 +34,11 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="AppBuilderTests.cs" />
-    <None Include="Loader\StartupLoaderTests.cs" />
+    <Compile Include="Loader\DoesNotFollowConvention.cs" />
+    <Compile Include="Loader\MultiConfigs.cs" />
+    <Compile Include="Loader\StartupLoaderTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Startup.cs" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
