+++ src/Tests/Gate.Builder.Tests/Loader/DoesNotFollowConvention.cs
@@ -1,0 +1,14 @@
+﻿using Gate.Owin;
+
+namespace DifferentNamespace
+{
+    public class DoesNotFollowConvention
+    {
+        public static int ConfigurationCalls;
+
+        public static void Configuration(IAppBuilder builder)
+        {
+            ConfigurationCalls += 1;
+        }
+    }
+}
