+++ src/Tests/Gate.Builder.Tests/Startup.cs
@@ -1,0 +1,15 @@
+﻿
+using Gate.Owin;
+
+namespace Gate.Builder.Tests
+{
+    public class Startup
+    {
+        public static void Configuration(IAppBuilder builder)
+        {
+            ++ConfigurationCalls;
+        }
+
+        public static int ConfigurationCalls { get; set; }
+    }
+}
